+++ ariles/tests/api_v2/regression_test_232.cpp
@@ -11,10 +11,6 @@
 #include "utility.h"
 
 
-#ifdef ARILES_VISITOR_yaml_cpp03
-#    include <ariles2/visitors/yaml_cpp03.h>
-#endif
-
 #ifdef ARILES_VISITOR_yaml_cpp
 #    include <ariles2/visitors/yaml_cpp.h>
 #endif