+++ ariles/tests/cmake/configs/options_conflict.cmake
@@ -1,5 +1,0 @@
-set(ARILES_CPP_STANDARD         "c++11" CACHE STRING "")
-
-set(ARILES_VISITOR_yaml_cpp03   "ON"    CACHE STRING "")
-set(ARILES_VISITOR_yaml_cpp     "ON"    CACHE STRING "")
-set(ARILES_VISITOR_rosparam     "OFF"   CACHE STRING "")