+++ ariles/tests/cmake/configs/options_deb_packages_bionic.cmake
@@ -7,5 +7,4 @@ set(ARILES_BUILD_REGRESSION_TESTS   "OFF"       CACHE STRING "")
 
 set(ARILES_VISITORS_DEFAULT_MODE    "ON"        CACHE STRING "")
 set(ARILES_VISITOR_rosparam         "melodic"   CACHE STRING "")
-set(ARILES_VISITOR_yaml_cpp03       "OFF"       CACHE STRING "")
 set(ARILES_VISITOR_msgpack          "OFF"       CACHE STRING "")