+++ ariles/tests/cmake/configs/options_deb_packages_xenial.cmake
@@ -4,9 +4,7 @@ set(ARILES_CPP_STANDARD             "c++11"     CACHE STRING "")
 
 set(ARILES_BUILD_REGRESSION_TESTS   "OFF"       CACHE STRING "")
 
-
 set(ARILES_VISITORS_DEFAULT_MODE    "ON"        CACHE STRING "")
 set(ARILES_VISITOR_rosparam         "kinetic"   CACHE STRING "")
 set(ARILES_VISITOR_jsonnet          "OFF"       CACHE STRING "")
-set(ARILES_VISITOR_yaml_cpp03       "OFF"       CACHE STRING "")
 set(ARILES_VISITOR_msgpack          "OFF"       CACHE STRING "")