+++ ariles/tests/cmake/configs/options_default.cmake
@@ -1,4 +1,2 @@
 set(ARILES_CPP_STANDARD         "c++11" CACHE STRING "")
 set(ARILES_CPP_SANITIZERS       "ON"    CACHE STRING "")
-
-set(ARILES_VISITOR_yaml_cpp03    "OFF"   CACHE STRING "")