+++ ariles/CHANGELOG.md
@@ -28,3 +28,7 @@ Removed
   definition.
 * `ARILES_DEFAULT_CONFIGURABLE_FLAGS` cmake parameter, can still be set using
   preprocessor. Corresponding test cases have also been dropped.
+* Ubuntu 14.04 (Trusty) tests.
+* yaml-cpp03 support.
+* C++03 tests.
+* cmake 2.x support.